Fixes Twilio responses.
Twilio expects an XML response to an incoming webhook, if it gets a content type like application/json it will show errors in the Twilio debugger. This commit ensures that the bot responds with a content type of text/xml and the an empty <Response></Response> element in the body.